Week13/Day2/ExerciseXP/requirements.txt
python-dotenv>=1.0.0
smolagents>=0.2.1
huggingface_hub>=0.22.0
